+++ pkg/api/clipboard_other.go
@@ -6,6 +6,10 @@ import "errors"
 
 func osClipboardFiles() []string { return nil }
 
+func osClipboardSeq() uint64 { return 0 }
+
+func osClipboardHasFiles() bool { return false }
+
 func osClipboardSetFiles([]string) error {
 	return errors.New("OS file clipboard is only supported on Windows")
 }